implementando flags de sincronização e fun~~ao de adicionar e remocer notas no databse local e remoto

+++ app/src/main/java/br/com/alura/ceep/repository/NotaRepository.kt
@@ -4,6 +4,7 @@ import br.com.alura.ceep.database.dao.NotaDao
 import br.com.alura.ceep.model.Nota
 import br.com.alura.ceep.webclient.NotaWebClient
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
 
 class NotaRepository(private val dao: NotaDao,
                      private val webClient: NotaWebClient) {
@@ -12,9 +13,13 @@ class NotaRepository(private val dao: NotaDao,
         return dao.buscaTodas()
     }
 
-    suspend fun atualizaTodas() {
+    private suspend fun atualizaTodas() {
         webClient.buscaTodas()?.let { notas ->
-            dao.salva(notas)
+           val notasQueSeraoSincronizadas =  notas.map {notaDaApi ->
+                notaDaApi.copy(sincronizada = true)
+
+            }
+            dao.salva(notasQueSeraoSincronizadas)
         }
     }
 
@@ -23,12 +28,31 @@ class NotaRepository(private val dao: NotaDao,
     }
 
     suspend fun remove(id: String) {
-        dao.remove(id)
+        dao.desativa(id)
+        if(webClient.remove(id)) {
+            dao.remove(id)
+        }
     }
 
     suspend fun salva(nota: Nota) {
         dao.salva(nota)
-        webClient.salva(nota)
+        if(webClient.salva(nota)){
+            val notaSincronizada = nota.copy(sincronizada = true)
+            dao.salva(notaSincronizada)
+
+        }
+    }
+
+    suspend fun sincroniza(){
+        val notasDesativadas = dao.buscaDesativadas().first()
+        notasDesativadas.forEach { notaDesativada ->
+            remove(notaDesativada.id)
+        }
+        val notaNaoSincronizada = dao.buscaNaoSincronizada().first()
+        notaNaoSincronizada.forEach{notasQueSeraoSincronizadas ->
+            salva(notasQueSeraoSincronizadas)
+        }
+        atualizaTodas()
     }
 
 }
